requirements.txt
numpy>=1.16.2
matplotlib>=3.0.2
pandas>=0.23.4
Pillow>=5.4.1
